requirements.txt
Django==5.0.6
django-ckeditor==6.7.1
Pillow==10.4.0